+++ climt/_components/__init__.py
@@ -1,5 +1,6 @@
 from .berger_solar_insolation import BergerSolarInsolation
 from .bucket_hydrology import BucketHydrology
+from .data_ocean import DataOcean
 from .dcmip import DcmipInitialConditions
 from .dry_convection import DryConvectiveAdjustment
 from .emanuel import (
@@ -9,9 +10,14 @@ from .emanuel import (
 from .grid_scale_condensation import GridScaleCondensation
 from .held_suarez import HeldSuarez
 from .instellation import Instellation
+from .land_ice import LandIce
+from .land_mask import LandMask
 from .cork import CorkLongwaveRadiation, CorkShortwaveRadiation
 from .radiation import Frierson06LongwaveOpticalDepth, GrayLongwaveRadiation
 from .rrtmg import RRTMGLongwave, RRTMGShortwave
+from .sea_ice import SeaIce
+from .second_best import SecondBEST
+from .simple_boundary_layer import SimpleBoundaryLayer
 from .simple_physics import SimplePhysics
 from .slab_surface import SlabSurface
 from .surface_ice import IceSheet
@@ -31,8 +37,14 @@ __all__ = (
     DcmipInitialConditions,
     IceSheet,
     Instellation,
+    LandMask,
     DryConvectiveAdjustment,
     BucketHydrology,
     CorkLongwaveRadiation,
     CorkShortwaveRadiation,
+    SecondBEST,
+    SeaIce,
+    LandIce,
+    DataOcean,
+    SimpleBoundaryLayer,
 )