+++ climt/_components/land_mask/__init__.py
@@ -1,0 +1,3 @@
+from .component import LandMask
+
+__all__ = (LandMask,)